2008-07-29  Kai Tietz  <kai.tietz@onevision.com>
        * misc/execv.c: Change return type to int.
        * misc/execve.c: Likewise.
        * misc/execvp.c: Likewise.
        * execvpe.c: Likewise.



git-svn-id: svn+ssh://svn.code.sf.net/p/mingw-w64/code/trunk@451 4407c894-4637-0410-b4f5-ada5f102cad1

+++ mingw-w64-crt/misc/execv.c
@@ -1,6 +1,6 @@
 #include <process.h>
 
-intptr_t __cdecl execv(const char *_Filename,char *const _ArgList[])
+int __cdecl execv(const char *_Filename,char *const _ArgList[])
 {
   return _execv (_Filename, (const char *const *)_ArgList);
 }
