+++ mingw-w64-crt/misc/execve.c
@@ -1,6 +1,6 @@
 #include <process.h>
 
-intptr_t __cdecl execve(const char *_Filename,char *const _ArgList[],char *const _Env[])
+int __cdecl execve(const char *_Filename,char *const _ArgList[],char *const _Env[])
 {
   return _execve (_Filename, (const char *const *)_ArgList, (const char * const *)_Env);
 }
